@@ -3,23 +3,23 @@ package gitfs
 import (
 	"os"
 
-	"github.com/go-git/go-billy/v5"
+	"github.com/go-git/go-git/v5"
 	"golang.org/x/net/webdav"
 )
 
 type Dir struct {
-	root billy.Filesystem
+	*git.Worktree
 	name string
 }
 
-func NewDir(root billy.Filesystem, name string) webdav.File {
-	return &Dir{root: root, name: name}
+func NewDir(worktree *git.Worktree, name string) webdav.File {
+	return &Dir{Worktree: worktree, name: name}
 }
 
 func (d *Dir) Close() error               { return nil }
 func (d *Dir) Read(p []byte) (int, error) { return 0, os.ErrInvalid }
 func (d *Dir) Readdir(int) ([]os.FileInfo, error) {
-	entries, err := d.root.ReadDir(d.name)
+	entries, err := d.Filesystem.ReadDir(d.name)
 	if err != nil {
 		return nil, err
 	}
@@ -40,7 +40,7 @@ func (d *Dir) Seek(offset int64, whence int) (int64, error) {
 }
 
 func (d *Dir) Stat() (os.FileInfo, error) {
-	return d.root.Stat(d.name)
+	return d.Filesystem.Stat(d.name)
 }
 
 func (d *Dir) Write(p []byte) (int, error) {
