@@ -4,6 +4,7 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/go-git/go-billy/v5"
 	"github.com/go-git/go-billy/v5/memfs"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/config"
@@ -42,11 +43,18 @@ func NewRepo(path string) *Repo {
 		panic(err)
 	}
 
+	head, err := repo.Head()
+	if err != nil {
+		panic(err)
+	}
+
 	return &Repo{
-		location:     location,
-		cwd:          path,
-		repo:         repo,
-		checkoutOpts: &git.CheckoutOptions{},
+		location: location,
+		cwd:      path,
+		repo:     repo,
+		checkoutOpts: &git.CheckoutOptions{
+			Branch: plumbing.ReferenceName(head.Name().String()),
+		},
 	}
 }
 
@@ -63,7 +71,7 @@ func (r *Repo) Cwd() string {
 	return r.cwd
 }
 
-func (r *Repo) HEAD() string {
+func (r *Repo) Head() string {
 	ref, err := r.repo.Head()
 	if err != nil {
 		panic(err)
@@ -85,13 +93,45 @@ func (r *Repo) SetBranch(branch string, force bool) {
 	}
 }
 
-func (r *Repo) GetDir() webdav.FileSystem {
+func (r *Repo) GetFileSystem() webdav.FileSystem {
 	worktree, err := r.repo.Worktree()
 	if err != nil {
 		panic(err)
 	}
 
-	worktree.Checkout(r.checkoutOpts)
+	err = worktree.Checkout(r.checkoutOpts)
+	if err != nil {
+		panic(err)
+	}
 
-	return NewFilesystem(worktree.Filesystem)
+	return NewFilesystem(worktree)
+}
+
+func (r *Repo) Walk(fn filepath.WalkFunc) error {
+	worktree, err := r.repo.Worktree()
+	if err != nil {
+		return err
+	}
+
+	walkFileInfo("/", worktree.Filesystem, fn)
+
+	return nil
+}
+
+func walkFileInfo(root string, fs billy.Filesystem, fn filepath.WalkFunc) {
+	file, err := fs.Lstat(root)
+	if err != nil {
+		fn(root, nil, err)
+	}
+	if !file.IsDir() {
+		fn(root, file, nil)
+		return
+	}
+	files, err := fs.ReadDir(root)
+	if err != nil {
+		fn(root, nil, err)
+	}
+	for _, file := range files {
+		walkFileInfo(filepath.Join(root, file.Name()), fs, fn)
+	}
 }
